+++ src/views/Me/Drafts/context.ts
@@ -1,0 +1,6 @@
+import { createCacheContext } from '~/components'
+
+import { MeDraftFeed_viewer_drafts_edges } from './__generated__/MeDraftFeed'
+
+export const DraftsContext =
+  createCacheContext<MeDraftFeed_viewer_drafts_edges[]>()
